Move filterFeatures to a separate file to be a util function
