+++ packages/plugins/Gfi/src/utils/filterFeatures.ts
@@ -1,0 +1,13 @@
+import { Feature as GeoJsonFeature } from 'geojson'
+import { FeaturesByLayerId } from '../types'
+
+export function filterFeatures(
+  featuresByLayerId: FeaturesByLayerId
+): Record<string, GeoJsonFeature[]> {
+  const entries = Object.entries(featuresByLayerId)
+  const filtered = entries.filter((keyValue) => Array.isArray(keyValue[1])) as [
+    string,
+    GeoJsonFeature[]
+  ][]
+  return Object.fromEntries(filtered)
+}
